[ObjC] rename the file to drop the "objectivec_" prefix.
Since "names.(h|cc)" now exists, drop the prefixing on all the files.
