+++ src/google/protobuf/compiler/main.cc
@@ -33,7 +33,7 @@
 #include "google/protobuf/compiler/java/kotlin_generator.h"
 #include "google/protobuf/compiler/command_line_interface.h"
 #include "google/protobuf/compiler/csharp/csharp_generator.h"
-#include "google/protobuf/compiler/objectivec/objectivec_generator.h"
+#include "google/protobuf/compiler/objectivec/generator.h"
 #include "google/protobuf/compiler/php/php_generator.h"
 #include "google/protobuf/compiler/python/generator.h"
 #include "google/protobuf/compiler/python/pyi_generator.h"
