+++ src/google/protobuf/compiler/objectivec/BUILD.bazel
@@ -21,7 +21,7 @@ cc_library(
     name = "names_internal",
     hdrs = [
         "names.h",
-        "objectivec_nsobject_methods.h",
+        "nsobject_methods.h",
     ],
     srcs = [
         "names.cc",
@@ -38,32 +38,32 @@ cc_library(
 cc_library(
     name = "objectivec",
     srcs = [
-        "objectivec_enum.cc",
-        "objectivec_enum_field.cc",
-        "objectivec_extension.cc",
-        "objectivec_field.cc",
-        "objectivec_file.cc",
-        "objectivec_generator.cc",
-        "objectivec_map_field.cc",
-        "objectivec_message.cc",
-        "objectivec_message_field.cc",
-        "objectivec_oneof.cc",
-        "objectivec_primitive_field.cc",
+        "enum.cc",
+        "enum_field.cc",
+        "extension.cc",
+        "field.cc",
+        "file.cc",
+        "generator.cc",
+        "map_field.cc",
+        "message.cc",
+        "message_field.cc",
+        "oneof.cc",
+        "primitive_field.cc",
     ],
     hdrs = [
-        "objectivec_enum.h",
-        "objectivec_enum_field.h",
-        "objectivec_extension.h",
-        "objectivec_field.h",
-        "objectivec_file.h",
-        "objectivec_generator.h",
-        "objectivec_helpers.h",
-        "objectivec_map_field.h",
-        "objectivec_message.h",
-        "objectivec_message_field.h",
-        "objectivec_oneof.h",
-        "objectivec_options.h",
-        "objectivec_primitive_field.h",
+        "enum.h",
+        "enum_field.h",
+        "extension.h",
+        "field.h",
+        "file.h",
+        "generator.h",
+        "helpers.h",
+        "map_field.h",
+        "message.h",
+        "message_field.h",
+        "oneof.h",
+        "options.h",
+        "primitive_field.h",
     ],
     copts = COPTS,
     include_prefix = "google/protobuf/compiler/objectivec",
@@ -81,7 +81,7 @@ cc_library(
 
 cc_test(
     name = "helpers_unittest",
-    srcs = ["objectivec_helpers_unittest.cc"],
+    srcs = ["helpers_unittest.cc"],
     deps = [
         ":objectivec",
         "//src/google/protobuf/io",
