+++ src/google/protobuf/compiler/objectivec/helpers_unittest.cc
@@ -28,7 +28,7 @@
 // (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-#include "google/protobuf/compiler/objectivec/objectivec_helpers.h"
+#include "google/protobuf/compiler/objectivec/helpers.h"
 #include "google/protobuf/io/zero_copy_stream_impl_lite.h"
 #include <gtest/gtest.h>
