+++ src/google/protobuf/compiler/objectivec/message.h
@@ -34,8 +34,8 @@
 #include <string>
 #include <set>
 #include <vector>
-#include "google/protobuf/compiler/objectivec/objectivec_field.h"
-#include "google/protobuf/compiler/objectivec/objectivec_oneof.h"
+#include "google/protobuf/compiler/objectivec/field.h"
+#include "google/protobuf/compiler/objectivec/oneof.h"
 #include "google/protobuf/descriptor.h"
 #include "google/protobuf/io/printer.h"
