+++ src/google/protobuf/compiler/objectivec/primitive_field.cc
@@ -28,13 +28,13 @@
 // (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-#include "google/protobuf/compiler/objectivec/objectivec_primitive_field.h"
+#include "google/protobuf/compiler/objectivec/primitive_field.h"
 
 #include <map>
 #include <string>
 
 #include "absl/strings/str_cat.h"
-#include "google/protobuf/compiler/objectivec/objectivec_helpers.h"
+#include "google/protobuf/compiler/objectivec/helpers.h"
 #include "google/protobuf/io/printer.h"
 
 namespace google {
@@ -71,7 +71,7 @@ const char* PrimitiveTypeName(const FieldDescriptor* descriptor) {
     case OBJECTIVECTYPE_ENUM:
       return "int32_t";
     case OBJECTIVECTYPE_MESSAGE:
-      return NULL;  // Messages go through objectivec_message_field.cc|h.
+      return NULL;  // Messages go through message_field.cc|h.
   }
 
   // Some compilers report reaching end of function even though all cases of
@@ -104,7 +104,7 @@ const char* PrimitiveArrayTypeName(const FieldDescriptor* descriptor) {
     case OBJECTIVECTYPE_ENUM:
       return "Enum";
     case OBJECTIVECTYPE_MESSAGE:
-      // Want NSArray (but goes through objectivec_message_field.cc|h).
+      // Want NSArray (but goes through message_field.cc|h).
       return "";
   }
